Added templates for feature configuration description/reference
The featureconfig.rst document should be used to describe the
configuration of the <XYZ> feature.  The content will be added to the
Brahmaputra configuration guide which intends to describe how to
configure the Brahmaputra release to a stable and ready state For
feature projects this should refer to the use of deployment tool
scenario selections and post deployment configuration specific to the
feature.  Also added is the postinstall.rst document for features to
describe how to check and analyze the result of test cases for
respective feature.

Change-Id: I3a2939cfcd733b1346bd9e459fbb70c855699bb8
Signed-off-by: Sofia Wallin <sofia.wallin@ericsson.com>