Update patch set 1
Patch Set 1: Code-Review-1

That doesn't work. I tried that "dirname $0" approach before and it does not work. The guidance I got from looking at this online was that the approach is fragile and based upon a hack that is unreliable. The passing of the directory, from a script that knows it, is the only reliable way to do this.

Patch-set: 1
Reviewer: Gerrit User 232 <232@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1