Update patch set 1
Patch Set 1:

Bryan must have removed the inactive Committers. I'm not familiar with those people, and I've been on the Copper project for about six months.  The TSC voted on 14 Feb 2017 to simplify the Committer removal process -- "A Committer who is disruptive, or has been inactive for an extended period (e.g., six or more months) may have his or her Committer status revoked by the Project Lead. The Project Lead is responsible for informing the TSC of any committers that are removed".   I don't know if Bryan informed the TSC or not.

Patch-set: 1